more code cleaning

+++ VividEngine/Font/VFont.cs
@@ -9,29 +9,29 @@ namespace Vivid3D.Font
         public string Path = "";
         public List<VGlyph> Glypth = new List<VGlyph>();
 
-        public int Width(string t)
+        public int Width ( string t )
         {
             int sw = 0;
-            foreach (char c in t)
+            foreach ( char c in t )
             {
-                VGlyph v = Glypth[(int)c];
-                sw += (int)((float)v.W / 1.3f);
+                VGlyph v = Glypth[c];
+                sw += ( int ) ( v.W / 1.3f );
             }
             return sw;
         }
 
-        public int Height()
+        public int Height ( )
         {
-            return Glypth[0].H;
+            return Glypth [ 0 ].H;
         }
 
-        public VFont(string path)
+        public VFont ( string path )
         {
             Path = path;
             FileStream fs = new FileStream(path, FileMode.Open, FileAccess.Read);
             BinaryReader r = new BinaryReader(fs);
 
-            for (int c = 0; c < 255; c++)
+            for ( int c = 0 ; c < 255 ; c++ )
             {
                 VGlyph g = new VGlyph
                 {
@@ -41,22 +41,22 @@ namespace Vivid3D.Font
 
                 byte[] img = new byte[g.W * g.H * 4];
 
-                for (int y = 0; y < g.H; y++)
+                for ( int y = 0 ; y < g.H ; y++ )
                 {
-                    for (int x = 0; x < g.W; x++)
+                    for ( int x = 0 ; x < g.W ; x++ )
                     {
                         int loc = (y * g.W * 4) + (x * 4);
-                        img[loc++] = r.ReadByte();
-                        img[loc++] = r.ReadByte();
-                        img[loc++] = r.ReadByte();
-                        img[loc] = r.ReadByte();
+                        img [ loc++ ] = r.ReadByte ( );
+                        img [ loc++ ] = r.ReadByte ( );
+                        img [ loc++ ] = r.ReadByte ( );
+                        img [ loc ] = r.ReadByte ( );
                     }
                 }
-                g.Img = new VTex2D(g.W, g.H, img, true);
-                Glypth.Add(g);
+                g.Img = new VTex2D ( g.W , g.H , img , true );
+                Glypth.Add ( g );
             }
 
-            fs.Close();
+            fs.Close ( );
             fs = null;
         }
     }
